mta: update stub for mta_hal_GetDHCPV6Info() to return dummy data instead of failure ( OFW-2106 )
OFW-2106 : Device.X_LGI-COM_MTA. and related parameters are not retrieved in GPV of "Device."

[Description]
When bulk get is performed, if get parameter value of a parameter
fails, failure is returned from there and remaining objects are not
retrieved.

[Solution]
Updated HAL to avoid reporting failure for mta_hal_GetDHCPV6Info(),
the API which provides data for:

  Device.X_CISCO_COM_MTA_V6.LeaseTimeRemaining

[Testing Performed]
GPV of 'Device.' from ACS and dmcli.

Signed-off-by: Dhanya Ravindran<dravindran.contractor@libertyglobal.com>